requirements.txt
aiogram>=3.0.0
requests==2.28.2
APScheduler==3.10.1
python-dotenv==1.0.0